refactor: extract mediaBlock to shared css-test-utils; fix 4-space indent in responsive layout test
